refactor(phase-3j): extract JiraFieldService
Phase 3j of ADR-002 god-class decomposition. Tenth slice of the
jira_client.py decomposition. Four field-metadata methods move from
jira_client.py into a new JiraFieldService exposed as self.fields.

Methods moved:
- get_field_metadata (public delegator)
- _get_field_metadata_via_createmeta (private helper, removed
  from JiraClient entirely; intra-cluster call from
  get_field_metadata stays inside the service)
- get_custom_fields (public delegator)
- get_issue_property (public delegator)

Inside the moved methods self.X becomes self._client.X and the
module-level logger becomes self._logger, matching the convention
used by the other Phase 3 services.

Numbers:
- jira_client.py: 1,158 -> 1,009 LOC (-149)
- jira_field_service.py: 0 -> 207 LOC (new)

Verification:
- pytest tests/unit: 953 passed
- mypy src/: clean (127 files)
- ruff check / ruff format: clean
